% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{bright_palette}
\alias{bright_palette}
\title{Color scales}
\usage{
bright_palette(primary = , other = , direction = 1)
}
\description{
This function is an custom color palette for ggplot2.
}
\details{
\describe{
\item{primary}{The primary color, default is Blue}
\item{other}{The other color that user can change, default is Red}
\item{direction}{Logical, 1 is default primary value is blue, other is red, then -1 is reverse}
\item{n}{The number of color selected}
...
}
}
\examples{
 #p + bright_palette(direction=1)(2)

}
