% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_bright_d}
\alias{scale_colour_bright_d}
\title{Palettes for discrete ggplot}
\usage{
scale_colour_bright_d(direction = )
}
\description{
This function gives you the color palettes to change the discrete value
}
\note{
The color include in this package are "Blue", "Green", "Red", "Yellow", "Purple", "Indigo", "Orange"
}
\examples{
  #p + scale_colour_bright_d(direction = 1)

}
