% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_bright_seq_c}
\alias{scale_colour_bright_seq_c}
\title{Add color for continuous variables}
\usage{
scale_colour_bright_seq_c(direction = , amount = )
}
\description{
This function gives you the color change from light to heavy when you want to compare the value.
}
\note{
The color include in this package are "Blue", "Green", "Red", "Yellow", "Purple", "Indigo", "Orange"
}
\examples{
  #p + scale_colour_bright_seq_c("Blue", direction = 1, amount = 0.3)

}
